refactgor: migrate from  js to ts with hold-event typing
