--- src/store/ls.js
@@ -1,7 +1,0 @@
-/*
- * @summary - Local Storage Module
- */
-
-export const state = () => ({})
-
-export const mutations = {}
