+++ src/store/ls.ts
@@ -1,0 +1,13 @@
+/*
+ * @summary - Local Storage Module
+ */
+
+interface LSState {
+  // Add state properties here as needed
+}
+
+export const state = (): LSState => ({})
+
+export const mutations = {
+  // Add mutations here as needed
+}
